+++ CHANGES.md
@@ -7,6 +7,7 @@ Changes in 7.1
 - Add Start Date input in Scheduler.php
 - Export (Excel) date to YYYY-MM-DD format (ISO) in Date.php & Preferences.php
 - Select Date Format: Add Preferences( 'DATE' ) in User.fnc.php, Preferences.php, Date.php & Side.php
+- Fix SQL error TITLE column limit to 50 characters in GradeLevels.php
 
 Changes in 7.0.3
 ----------------